requirements.txt
Flask
python-dotenv
google-generativeai
youtube-transcript-api
psycopg2-binary
